Tweak Makefile to build doomgeneric the Holodeck's way
Tweak Makefile to build doomgeneric like 32-bit programs in Holodeck. We
link with our standard static library, and try to compile the code.

For now, when I add each error I get from the Makefile to out libc,
compile again Holodeck to rebuild the libc with the new function
included, and copy libc to doomgeneric.

Signed-off-by: Dimitris Charisis <dchar@cslab.ece.ntua.gr>