Complete first stage of doomgeneric port
* Add `crt.asm` to properly link the final executable. If we don't
  link with our minimal runtime system, [crt.o] the linker throws a
  warning that it cannot find the `_start` function. This function
  is the initial point that calls main and it is defined in our
  `crt.asm` function. So add this function, and fix Makefile to take
  it into account when building doomgeneric executable.

* Add `doomgeneric_hdeck.c` file. This is our core implementation for
  doomgeneric port. For now it does not implement anything.

At this stage, if you compile and run the produced `doomgeneric`
executable in Holodock, via a command like:
`./hdeck --32 doomgeneric --fb shared`, it prints a message `Doom
Generic 0.1` and crashes!
This is great. Indeed this message is called from
`I_PrintBanner(PACKAGE_STRING)` in `D_DoomMain()` function , and you
can trace the calls and see the functions being called.

Note: `I_PrintBanner()` calls `putchar()` and `puts()` which are
implemented in Holodeck's libc and this is why we can see the output!

This commit ends this branch, the dependencies are resolved [most of
them in `doomgdeps.c` in Holodeck's libc and the rest of them in
`doomgeneric_hdeck.c`].

Signed-off-by: Dimitris Charisis <dchar@cslab.ece.ntua.gr>